Add typing for topojson-server
