+++ types/topojson-server/tsconfig.json
@@ -2,8 +2,7 @@
     "compilerOptions": {
         "module": "commonjs",
         "lib": [
-            "es6",
-            "dom"
+            "es6"
         ],
         "noImplicitAny": true,
         "noImplicitThis": true,
@@ -19,7 +18,6 @@
     },
     "files": [
         "index.d.ts",
-        "topojson-tests.ts",
-        "test/atlas-tests.ts"
+        "topojson-server-tests.ts"
     ]
 }
